NLP_LSTM_CNN/requirements.gpu.txt
python-dotenv==0.1.0
tweepy==3.5.0
tensorflow-gpu==1.14.0
pandas == 1.0.1
nltk == 3.4.5
scikit_learn == 0.23.2
emoji == 1.2.0
seaborn==0.9.0
tqdm==4.32.2
matplotlib == 3.1.3
numpy == 1.18.1